Unify duccstats component into one
Reusable for killers or friends
